requirements.txt

cx-Oracle==8.3.0
Flask==3.1.0
Flask-Cors==5.0.0
huggingface-hub==0.27.1
inflect==7.5.0
pandas==2.2.3
python-dotenv==1.0.1
RapidFuzz==3.11.0
tokenizers==0.21.0
transformers==4.48.0
typing_extensions==4.12.2
uvicorn==0.34.0
torch